Setting up automatic deploys from Travis to Github Pages (#43)
After my journey through 3 PRs (#40, #41 and #42) I finally made it to this point. After @HHogg's comment I updated everything again and now we are able to push to the gh-pages branch with gh-pages.

The basis of this deployment is the usage of deploy keys instead of Github OAuth tokens. This makes the setup user independent and instead ties it to the repository. The deploy key is stored in the repository settings and is encrypted with the Travis CLI. The encryption keys are stored in the Travis settings.

As written in #42 deployments with the deploy key are only possible from a branch of the repository. It won't run on pull requests from forks. (That's why ssh.sh and deploy.sh both have a condition to exit in case it is a PR or another branch except master.

Long-term the condition probably should be updated to only run when a tag was added instead of running whenever there is a commit to master