feat(BarChart): highlight row when mouse is anywhere inside the row
When you hovered over the benchmark line other rows have lost their fade
state.

The root cause for this is that the selected row is set when you hover
over a bar and the benchmark line is not part of the bar.

This PR fixes this by introducing a separate state, when you hover over
the over a row, by just setting `selectedIndex`. When you hover over
over a bar the `selectedColor` then set.

This made it possible to keep the other rows faded, when you hovered
over the benchmark line.