feat(TextInput): Removed condition preventing "clear" icon AND input icon
BREAKING CHANGE: This is going to cause some design "glitches" where a TextInput
is used with the "onClear" property AND a TextInputIcon as a child WITHOUT the
"align" property being set to "left".

The following code examples are affected:

TextInputIcon with no "align" property:
In this case previously the "TextInputIcon" wouldn't be visible but with this
change is going to be visible and aligned on the right.
```js
<TextInput onClear={ () => {} }>
  <TextInputIcon name="magnify-glass" />
</TextInput>
```

You can fix this by adding the "align" property to the TextInputIcon
```js
<TextInput onClear={ () => {} }>
  <TextInputIcon name="magnify-glass" align="left" />
</TextInput>
```

Another affected example would be this:
Same as in the previous example the "TextInputIcon" wouldn't be visible but with
this change it would be visible and aligned on the right.
```js
<TextInput onClear={ () => {} }>
  <TextInputIcon name="magnify-glass" align="right" />
</TextInput>
```

Same as before you can fix this by changing the "align" property on the
"TextInputIcon" to "left":
```js
<TextInput onClear={ () => {} }>
  <TextInputIcon name="magnify-glass" align="left" />
</TextInput>
```