fix: 🐛 Ensure onRequestClose is passed to dialog header
The prevoius method of using DisplayName to clone in onRequest close to
the dialog header didn't work in production because displayName was
getting removed in the build process, not sure by what. Using context
should be more relible and also have the benifit of allowing dialog
header to be wrapped in other elements.